LayoutTests/platform/ios-simulator/accessibility/table-cell-for-row-col-expected.txt

This test makes sure that you can retrieve elements within a cell by their row and column

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cell2.cellForColumnAndRow(2, 0).isEqual(header) is true
PASS header.cellForColumnAndRow(1, 2).isEqual(cell6) is true
PASS successfullyParsed is true

TEST COMPLETE
header 1	header 2	header 3
cell 1	
cell
2
cell 4	cell 5	cell
6
